fruit_store/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
django-allauth==0.63.1
idna==3.7
mysqlclient==2.2.4
oauthlib==3.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
razorpay==1.4.2
requests==2.32.0
requests-oauthlib==2.0.0
setuptools==70.0.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
